YouBuddy/src/requirements.txt
google-adk[all]
google-cloud-secret-manager>=2.16.0,<3.0.0
google-auth>=2.22.0,<3.0.0